merge genchangelog to repoman server
In the past, we only had genchangelog on the server and repoman on the
client. Then more stuff was moved to the server, and that repoman server
was introduced. These days repoman only calls repoman server, so
genchangelog is internal API, as a result it doesn't make sense to have
it as a separate script anymore.